shell-ui: Add notificationcenter in the shell context and Expose the useNotificationCenter from shell
